Accounts + Backend Rework (#404)
* made wakeTimeHour to localtime to support minutes

* chore: update openapi types

* created daos

* impl jwt in pom

* implemented dtos, daos and db logic

* auth endpoints working

* implemented in spring security

* implemented GET /groups/user

* made endpoint name static

* added leave group endpoint

* renamed leave endpoint

* implemented createdBy

* auth for every /groups path. excluding some endpoints

* implemented group member dto + creating group member on group creation

* add createdBy to rules

* copying createdBy from old season

* saving createdBy for seasons

* saving for matches

* using seasondto instead of season dao

* use many-to-one relation

* soft delete group members

* join group endpoint

* fixed tests + made auth controller post endpoints

* loading jwt secret via env variables

* fix: add JWT_TOKEN to pass tests in ci

* fix

* fix: this time

* lsot

* chore: update openapi types

* added todo

* add http request asserts

* better create test

* more create tests

* more tests

* moreee tests

* group join tests

* group tests complete

* fixed group tests

* added createTestGroup methods

* added requestutils + start season tests

* season tests

* changed season update to use dto

* fixed incomplete season update dto

* impl new waketime in tests and code

* removed some todos

* finished season tests

* testing for valid userId in created + rule move tests

* changed rule move creation

* rule move tests

* added test for creation of rule moves on new season

* dont create beerpong rules for other presets

* rule copy test

* rule tests

* profiles test

* profile tests

* player tests

* player delete tests

* test copying of players on season start

* player tests

* base of match tests

* create tests

* fix tests

* force team amount

* invalid create tests

* fix tests

* remove todos

* more match tests

* match overview test

* overview invalid tests

* overview all tests

* remove old match test

* group test descr

* add team size check to match update

* match update invalid tests

* delete test

* invalid match delete tests

* match update test

* leaderboard tests

* remove useless boolean

* first leaderboard tests

* fix tests

* fix merge

* add group tests

* fix tests

* disabled elo test in normal test runtime

* feat(frontend): accounts

* devops: force openapi action to rerun

* chore: update openapi types

* feat(frontend): auth

* devops: re-disable openapi action

* better application properties

* new api

* daos

* test

* all daos and repos

* base of auth

* assets

* moved dtos + group presets

* rules and rule moves

* match moves

* players + profile start

* profiles

* groups

* teammembers

* teams

* matches

* seasons

* leaderboard service

* season service

* controller

* controller2

* conroller3

* compile

* tests

* fixed first tests

* player tests working

* first match test working

* fix matchdto validation

* match tests working nearly

* tests working

* fix inconsistency in profile test

* remove debug

* removed some todos

* fix ci

* change ci env

* ... grrrr

* fix ci

* jetzt aber

* Merge staging into feat/accounts-new (#402)

* Bump app version to 1.0.92 (#400)


* Bump app version to 1.0.93 (#401)


---------


* fixed auth flow

* fixed asset deleting

* fix asset deletion

* giving asset upload out

* validate existingTeamIds

* replaced api module wiht api-new module

* cleanup code

* fix n+1 query

* docker image working

* changed envs

* regen openapi

* testing smth

* fix builds

---------

+++ api/src/main/java/pro/beerpong/api/service/AssetService.java
@@ -9,13 +9,15 @@ import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.context.request.ServletRequestAttributes;
 import pro.beerpong.api.mapping.AssetMapper;
 import pro.beerpong.api.model.dao.Asset;
-import pro.beerpong.api.model.dto.AssetCropDto;
-import pro.beerpong.api.model.dto.AssetMetadataDto;
-import pro.beerpong.api.model.dto.AssetUploadResponse;
+import pro.beerpong.api.model.dto.assets.AssetCropDto;
+import pro.beerpong.api.model.dto.assets.AssetMetadataDto;
+import pro.beerpong.api.model.dto.assets.AssetUploadResponse;
 import pro.beerpong.api.repository.AssetRepository;
 import pro.beerpong.api.util.AssetType;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.model.DeleteObjectRequest;
+import software.amazon.awssdk.services.s3.model.HeadObjectRequest;
+import software.amazon.awssdk.services.s3.model.NoSuchKeyException;
 import software.amazon.awssdk.services.s3.model.PutObjectRequest;
 import software.amazon.awssdk.services.s3.presigner.S3Presigner;
 
@@ -36,49 +38,63 @@ public class AssetService {
         return assetRepository.existsById(assetId);
     }
 
-    public void deleteAsset(String assetId) {
-        client.deleteObject(DeleteObjectRequest.builder()
-                .bucket(bucket)
-                .key(assetId)
-                .build());
-
-        assetRepository.deleteById(assetId);
-    }
-
     public AssetMetadataDto getAssetData(String assetId) {
-        return assetMapper.assetToAssetMetadataDto(assetRepository.findById(assetId).orElse(null));
+        return assetRepository.findById(assetId)
+                .map(assetMapper::assetToAssetMetadataDto)
+                .orElse(null);
     }
 
     public Asset map(AssetMetadataDto dto) {
         return assetMapper.assetMetadataDtoToAsset(dto);
     }
 
-    public AssetUploadResponse storeAsset(AssetType assetType, @Nullable AssetCropDto assetCropDto) {
+    public Asset storeAsset(AssetType assetType, @Nullable AssetCropDto assetCropDto) {
         if (assetCropDto != null) {
-            return this.storeAsset(assetType, assetCropDto.getOffsetX(), assetCropDto.getOffsetY(),
-                    assetCropDto.getZoom());
+            return this.storeAsset(
+                    assetType,
+                    assetCropDto.getOffsetX(),
+                    assetCropDto.getOffsetY(),
+                    assetCropDto.getZoom()
+            );
         } else {
             return this.storeAsset(assetType);
         }
     }
 
-    public AssetUploadResponse storeAsset(AssetType assetType) {
+    public Asset storeAsset(AssetType assetType) {
         return this.storeAsset(assetType, 0.0D, 0.0D, 0.0D);
     }
 
-    public AssetUploadResponse storeAsset(AssetType assetType, double offsetX, double offsetY, double zoom) {
-        var asset = new Asset();
-        asset.setType(assetType);
-        asset.setOffsetX(offsetX);
-        asset.setOffsetY(offsetY);
-        asset.setZoom(zoom);
+    public Asset storeAsset(AssetType assetType, double offsetX, double offsetY, double zoom) {
+        var asset = new Asset(
+                null,
+                assetType,
+                offsetX,
+                offsetY,
+                zoom
+        );
 
         asset = assetRepository.save(asset);
 
-        return createPutUpload(assetRepository.save(asset), resolveImageContentType());
+        return asset;
     }
 
-    public AssetUploadResponse createPutUpload(Asset asset, String contentType) {
+    public AssetUploadResponse uploadAsset(Asset asset) {
+        return createPutUpload(asset, resolveImageContentType());
+    }
+
+    public void deleteAsset(String assetId) {
+        if (existsInBucket(bucket, assetId)) {
+            client.deleteObject(DeleteObjectRequest.builder()
+                    .bucket(bucket)
+                    .key(assetId)
+                    .build());
+        }
+
+        assetRepository.deleteById(assetId);
+    }
+
+    private AssetUploadResponse createPutUpload(Asset asset, String contentType) {
         var putReq = PutObjectRequest.builder()
                 .bucket(bucket)
                 .key(asset.getId())
@@ -88,17 +104,9 @@ public class AssetService {
         var presigned = presigner.presignPutObject(b -> b
                 .signatureDuration(Duration.ofMinutes(5))
                 .putObjectRequest(putReq));
+        var response = assetMapper.assetToUploadResponse(asset);
 
-        var response = new AssetUploadResponse();
-        response.setId(asset.getId());
-        response.setUrl(assetMapper.generateUrl(asset));
         response.setSingleUploadUrl(presigned.url().toString());
-
-        response.setZoom(asset.getZoom());
-        response.setOffsetX(asset.getOffsetX());
-        response.setOffsetY(asset.getOffsetY());
-
-        response.setType(asset.getType());
 
         return response;
     }
@@ -115,5 +123,17 @@ public class AssetService {
             }
         }
         return "image/png";
+    }
+
+    private boolean existsInBucket(String bucket, String key) {
+        try {
+            client.headObject(HeadObjectRequest.builder()
+                    .bucket(bucket)
+                    .key(key)
+                    .build());
+            return true;
+        } catch (NoSuchKeyException e) {
+            return false;
+        }
     }
 }
